fix: rename hook/lib scripts .js → .cjs for ESM project compatibility
Projects with "type": "module" in package.json treat .js files as ESM,
breaking our hooks that use require(). Renaming to .cjs forces Node to
treat them as CommonJS regardless of the target project's module type.

- Renamed 21 files: scripts/hooks/*.js → *.cjs, scripts/lib/*.js → *.cjs,
  skills/strategic-compact/scripts/suggest-compact.js → .cjs
- Updated 16 inter-file require() paths to include .cjs extension
- Updated 16 hook command references in hooks.json
