+++ hooks/hooks.json
@@ -7,7 +7,7 @@
         "hooks": [
           {
             "type": "command",
-            "command": "node \"${CLAUDE_CONFIG_DIR}/scripts/hooks/guardian.js\""
+            "command": "node \"${CLAUDE_CONFIG_DIR}/scripts/hooks/guardian.cjs\""
           }
         ],
         "description": "Guardian: evaluate tool call risk and block dangerous operations"
@@ -17,7 +17,7 @@
         "hooks": [
           {
             "type": "command",
-            "command": "node \"${CLAUDE_CONFIG_DIR}/scripts/hooks/pr-intent-gate.js\""
+            "command": "node \"${CLAUDE_CONFIG_DIR}/scripts/hooks/pr-intent-gate.cjs\""
           }
         ],
         "description": "PR Intent Gate: require ## Why section in PR body"
@@ -27,7 +27,7 @@
         "hooks": [
           {
             "type": "command",
-            "command": "node \"${CLAUDE_CONFIG_DIR}/scripts/hooks/merged-branch-guard.js\""
+            "command": "node \"${CLAUDE_CONFIG_DIR}/scripts/hooks/merged-branch-guard.cjs\""
           }
         ],
         "description": "Merged Branch Guard: block operations on branches with merged PRs"
@@ -37,7 +37,7 @@
         "hooks": [
           {
             "type": "command",
-            "command": "node \"${CLAUDE_CONFIG_DIR}/scripts/hooks/review-gate.js\""
+            "command": "node \"${CLAUDE_CONFIG_DIR}/scripts/hooks/review-gate.cjs\""
           }
         ],
         "description": "Review gate: require code review before commits"
@@ -57,7 +57,7 @@
         "hooks": [
           {
             "type": "command",
-            "command": "node \"${CLAUDE_CONFIG_DIR}/scripts/hooks/skill-usage-tracker.js\""
+            "command": "node \"${CLAUDE_CONFIG_DIR}/scripts/hooks/skill-usage-tracker.cjs\""
           }
         ],
         "description": "Track skill invocations for usage analytics"
@@ -67,7 +67,7 @@
         "hooks": [
           {
             "type": "command",
-            "command": "node \"${CLAUDE_CONFIG_DIR}/skills/strategic-compact/scripts/suggest-compact.js\""
+            "command": "node \"${CLAUDE_CONFIG_DIR}/skills/strategic-compact/scripts/suggest-compact.cjs\""
           }
         ],
         "description": "Suggest manual compaction at logical intervals"
@@ -79,7 +79,7 @@
         "hooks": [
           {
             "type": "command",
-            "command": "node \"${CLAUDE_CONFIG_DIR}/scripts/hooks/pre-compact.js\""
+            "command": "node \"${CLAUDE_CONFIG_DIR}/scripts/hooks/pre-compact.cjs\""
           }
         ],
         "description": "Save state before context compaction"
@@ -91,7 +91,7 @@
         "hooks": [
           {
             "type": "command",
-            "command": "node \"${CLAUDE_CONFIG_DIR}/scripts/hooks/session-start.js\""
+            "command": "node \"${CLAUDE_CONFIG_DIR}/scripts/hooks/session-start.cjs\""
           }
         ],
         "description": "Load previous context and detect package manager on new session"
@@ -103,7 +103,7 @@
         "hooks": [
           {
             "type": "command",
-            "command": "node \"${CLAUDE_CONFIG_DIR}/scripts/hooks/pr-url-logger.js\""
+            "command": "node \"${CLAUDE_CONFIG_DIR}/scripts/hooks/pr-url-logger.cjs\""
           }
         ],
         "description": "Log PR URL and provide review command after PR creation"
@@ -113,7 +113,7 @@
         "hooks": [
           {
             "type": "command",
-            "command": "node \"${CLAUDE_CONFIG_DIR}/scripts/hooks/ci-monitor.js\"",
+            "command": "node \"${CLAUDE_CONFIG_DIR}/scripts/hooks/ci-monitor.cjs\"",
             "async": true,
             "timeout": 30
           }
@@ -125,7 +125,7 @@
         "hooks": [
           {
             "type": "command",
-            "command": "node \"${CLAUDE_CONFIG_DIR}/scripts/hooks/prettier-format.js\""
+            "command": "node \"${CLAUDE_CONFIG_DIR}/scripts/hooks/prettier-format.cjs\""
           }
         ],
         "description": "Auto-format JS/TS files with Prettier after edits"
@@ -135,7 +135,7 @@
         "hooks": [
           {
             "type": "command",
-            "command": "node \"${CLAUDE_CONFIG_DIR}/scripts/hooks/typescript-check.js\""
+            "command": "node \"${CLAUDE_CONFIG_DIR}/scripts/hooks/typescript-check.cjs\""
           }
         ],
         "description": "TypeScript check after editing .ts/.tsx files"
@@ -145,7 +145,7 @@
         "hooks": [
           {
             "type": "command",
-            "command": "node \"${CLAUDE_CONFIG_DIR}/scripts/hooks/console-log-warn.js\""
+            "command": "node \"${CLAUDE_CONFIG_DIR}/scripts/hooks/console-log-warn.cjs\""
           }
         ],
         "description": "Warn about console.log statements after edits"
@@ -157,7 +157,7 @@
         "hooks": [
           {
             "type": "command",
-            "command": "node \"${CLAUDE_CONFIG_DIR}/scripts/hooks/check-console-log.js\""
+            "command": "node \"${CLAUDE_CONFIG_DIR}/scripts/hooks/check-console-log.cjs\""
           }
         ],
         "description": "Check for console.log in modified files after each response"
@@ -169,7 +169,7 @@
         "hooks": [
           {
             "type": "command",
-            "command": "node \"${CLAUDE_CONFIG_DIR}/scripts/hooks/session-end.js\""
+            "command": "node \"${CLAUDE_CONFIG_DIR}/scripts/hooks/session-end.cjs\""
           }
         ],
         "description": "Persist session state on end"
@@ -179,7 +179,7 @@
         "hooks": [
           {
             "type": "command",
-            "command": "node \"${CLAUDE_CONFIG_DIR}/scripts/hooks/evaluate-session.js\""
+            "command": "node \"${CLAUDE_CONFIG_DIR}/scripts/hooks/evaluate-session.cjs\""
           }
         ],
         "description": "Evaluate session for extractable patterns"
