+++ scripts/hooks/check-console-log.cjs
