+++ scripts/hooks/console-log-warn.cjs
