+++ scripts/hooks/pr-intent-gate.cjs
@@ -6,7 +6,7 @@
  * Fail-open: if anything goes wrong, exits 0 (allows).
  */
 
-const { readStdinJson, log } = require('../lib/utils');
+const { readStdinJson, log } = require('../lib/utils.cjs');
 
 async function main() {
   const input = await readStdinJson();
