+++ scripts/hooks/prettier-format.cjs
