+++ scripts/hooks/session-end.cjs
@@ -21,7 +21,7 @@ const {
   findSpecsDir,
   getGitModifiedFiles,
   log
-} = require('../lib/utils');
+} = require('../lib/utils.cjs');
 
 async function main() {
   const sessionsDir = getSessionsDir();
