+++ scripts/hooks/typescript-check.cjs
