+++ skills/strategic-compact/scripts/suggest-compact.cjs
